move usb descriptors into usb.h
